A good method to solve this problem is to create an array arr which has a size equal to the largest value of right. In this case, it would be max(trip[2]) for all trip in trips. Then, iterate over input - for each value, in this case numPassengers, increment arr[left] (in this case arr[from]) by the value, and decrement arr[right] (in this case arr[to]) by the value.
Essentially, what we are doing is we are building the "road" with an array. For each point on the road, we are marking the net change of people in the car. Now, we can walk through the array with a prefix sum, which represents how many people are in the car at the moment. If this value ever exceeds capacity, then we can't complete all the trips.